Generic two-stage designs for single-arm trials with binary endpoint

[todo]

Design(n2, c2, label = "")

sample_size(design, x1 = NULL)

critical_value(design, x1)

futility_region(design)

efficacy_region(design)

continuation_region(design)

pmf(design, x1, x2, p)

probability_to_reject(design, p, x1 = NULL)

expected_sample_size(design, p)

sample_space(design)

reject(x1, x2, design)

# S3 method for Design
as_tibble(x, ...)

# S3 method for Design
plot(x, tbl_power_annotations = NULL, ...)

plot_designs(
  ...,
  textsize = 1.75,
  ystep = 10,
  yexpandlower = 0.1,
  skip_listing = NULL
)

plot_power(
  ...,
  annotations = NULL,
  min.segment.length = Inf,
  annotation_text_size = 1.5
)

plot_ess(
  ...,
  annotations = NULL,
  min.segment.length = Inf,
  annotation_text_size = 2
)

plot_compare_designs(
  ...,
  annotations = NULL,
  min.segment.length = Inf,
  annotation_text_size = 2,
  args_design_plot = list(textsize = 1.75, ystep = 10),
  args_power_plot = list(annotations = annotations, min.segment.length =
    min.segment.length, annotation_text_size = annotation_text_size),
  args_ess_plot = list(annotations = annotations, min.segment.length =
    min.segment.length, annotation_text_size = annotation_text_size)
)

Arguments

n2
numeric vector with stage-two sample sizes for x1 = 0, 1, ..., n1

c2
numeric vectot with stage-two decision boundary for x1 = 0, 1, ..., n1; the design rejects H0: p <= p0 iff X2 > c_2(X1); use +/- Inf for early futility/efficacy (n2 = 0).

label
an optional label to use for pretty plotting etc.

design
a Design object

x1
stage-one number of results

x2
stage-two number of responses

p
response probability

...
one more more Design objects to plot

textsize
size of c2/n2 annotation in design plot

ystep
tick marks step in design plot y axis

yexpandlower
factor for (lower) y axis expansion in design plot,

skip_listing
character vector with design labels for which to not include the c2/n2 listings

annotations
vector of response prbabilities to annotate

min.segment.length
minimal lenght of arrow to annotations

annotation_text_size
annotation text size

args_design_plot
list of arguments for the design plot

args_power_plot
list of arguments for the power plot

args_ess_plotlist
of arguments for the expected sample size plot